+++ app/helpers/custom_field_properties_helper.rb
@@ -1,6 +1,6 @@
 module CustomFieldPropertiesHelper
   def custom_field_properties_edit_link(custom_field_property)
-    if custom_field_editable?(@custom_field)
+    if custom_field_editable?(@custom_field) && (client_custom_field? ? policy(@custom_formable).edit? : true)
       link_to edit_polymorphic_path([@custom_formable, custom_field_property], custom_field_id: @custom_field) do
         content_tag :div, class: 'btn btn-outline btn-success' do
           fa_icon('pencil')
@@ -16,7 +16,7 @@ module CustomFieldPropertiesHelper
   end
 
   def custom_field_properties_destroy_link(custom_field_property)
-    if custom_field_editable?(@custom_field)
+    if custom_field_editable?(@custom_field) && (client_custom_field? ? policy(@custom_formable).destroy? : true)
       link_to polymorphic_path([@custom_formable, custom_field_property], custom_field_id: @custom_field), method: :delete, data: { confirm: t('are_you_sure') } do
         content_tag :div, class: 'btn btn-outline btn-danger' do
           fa_icon('trash')
@@ -32,7 +32,7 @@ module CustomFieldPropertiesHelper
   end
 
   def custom_field_properties_new_link
-    if custom_field_editable?(@custom_field)
+    if custom_field_editable?(@custom_field) && (client_custom_field? ? policy(@custom_formable).new? : true)
       link_to new_polymorphic_path([@custom_formable, CustomFieldProperty], custom_field_id: @custom_field.id) do
         content_tag :div, class: 'btn btn-outline btn-primary form-btn' do
           "#{t('.add_new')} #{@custom_field.form_title}"
@@ -45,5 +45,9 @@ module CustomFieldPropertiesHelper
         end
       end
     end
+  end
+
+  def client_custom_field?
+    @custom_formable.class.name == 'Client'
   end
 end
